requirements.txt
python-dotenv
elsapy
pandas

ratelimit